move setup to a component to get refresh on toggle change
