+++ src/app/system/devTools/DevMenu/Components/DevTools.tsx
@@ -10,7 +10,7 @@ import { DevToggleName, devToggles } from "app/store/config/features"
 import { Versions } from "app/store/migration"
 import { DevMenuButtonItem } from "app/system/devTools/DevMenu/Components/DevMenuButtonItem"
 import { DevToggleItem } from "app/system/devTools/DevMenu/Components/DevToggleItem"
-import { eigenSentryReleaseName } from "app/system/errorReporting/sentrySetup"
+import { eigenSentryReleaseName } from "app/system/errorReporting/setupSentry"
 import { dismissModal, navigate } from "app/system/navigation/navigate"
 import { RelayCache } from "app/system/relay/RelayCache"
 import { useUnleashEnvironment } from "app/utils/experiments/hooks"
